+++ src/components/Sidebar.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import Link from 'next/link';
 import SidebarItems from './SidebarItems';
-import { Avatar, AvatarFallback } from './ui/avatar';
 
 import { AuthSession, getUserAuth } from '@/lib/auth/utils';
+
+import { Avatar, AvatarFallback } from '@/components/ui';
 
 const Sidebar = async () => {
   const session = await getUserAuth();
